Separate part of InfoboxBuilderPage into Spinner component
