+++ src/test/java/com/wikia/webdriver/testcases/infoboxbuilder/InfoboxBuilderTests.java
@@ -6,6 +6,7 @@ import com.wikia.webdriver.common.core.annotations.Execute;
 import com.wikia.webdriver.common.core.helpers.User;
 import com.wikia.webdriver.common.templates.NewTestTemplate;
 import com.wikia.webdriver.elements.mercury.components.Sidebar;
+import com.wikia.webdriver.elements.mercury.components.Spinner;
 import com.wikia.webdriver.elements.mercury.components.Subhead;
 import com.wikia.webdriver.elements.mercury.pages.InfoboxBuilderPage;
 import com.wikia.webdriver.elements.oasis.pages.TemplateEditPage;
@@ -67,7 +68,7 @@ public class InfoboxBuilderTests extends NewTestTemplate {
     builderSidebar.addRowComponent();
     new Subhead().clickPublish();
 
-    Assertion.assertTrue(builderPage.isSpinnerPresent());
+    Assertion.assertTrue(new Spinner().isSpinnerPresent());
 
     Assertion.assertEquals("infoboxbuildersavingtemplate",
                            new TemplatePage().getHeaderText().toLowerCase());
@@ -168,7 +169,7 @@ public class InfoboxBuilderTests extends NewTestTemplate {
 
     new Subhead().clickPublish();
 
-    Assertion.assertTrue(builderPage.isSpinnerPresent());
+    Assertion.assertTrue(new Spinner().isSpinnerPresent());
     Assertion.assertTrue(template.isTemplatePagePresent());
 
     String infoboxTitle = template.getPortableInfobox().getTitleTextWithIndex(0);
@@ -190,7 +191,7 @@ public class InfoboxBuilderTests extends NewTestTemplate {
     builderSidebar.setTitleToUseArticleName();
     new Subhead().clickPublish();
 
-    Assertion.assertTrue(builderPage.isSpinnerPresent());
+    Assertion.assertTrue(new Spinner().isSpinnerPresent());
     Assertion.assertTrue(template.isTemplatePagePresent());
 
     String infoboxTitle = template.getPortableInfobox().getTitleTextWithIndex(0);
